+++ semcore/ellipsis/src/Ellipsis.tsx
@@ -91,6 +91,16 @@ const defaultTooltipProps = [
   'cursorAnchoring',
 ];
 
+const setFontSettings = (element: HTMLElement, styleElement: CSSStyleDeclaration): void => {
+  element.style.fontFamily = styleElement.getPropertyValue('font-family');
+  element.style.fontSize = styleElement.getPropertyValue('font-size');
+  element.style.fontWeight = styleElement.getPropertyValue('font-weight');
+  element.style.lineHeight = styleElement.getPropertyValue('line-height');
+  element.style.fontFeatureSettings =
+          styleElement.getPropertyValue('font-feature-settings');
+  element.style.fontVariantNumeric = styleElement.getPropertyValue('font-variant-numeric');
+};
+
 const createMeasurerElement = (element: HTMLDivElement, text?: string) => {
   const styleElement = window.getComputedStyle(element, null);
   const temporaryElement = document.createElement('temporary-block');
@@ -100,16 +110,10 @@ const createMeasurerElement = (element: HTMLDivElement, text?: string) => {
   temporaryElement.style.right = '0%';
   temporaryElement.style.bottom = '0%';
   temporaryElement.style.visibility = 'hidden';
-  temporaryElement.style.fontFamily = styleElement.getPropertyValue('font-family');
-  temporaryElement.style.fontSize = styleElement.getPropertyValue('font-size');
-  temporaryElement.style.fontWeight = styleElement.getPropertyValue('font-weight');
-  temporaryElement.style.lineHeight = styleElement.getPropertyValue('line-height');
   temporaryElement.style.whiteSpace = styleElement.getPropertyValue('white-space');
   temporaryElement.style.wordWrap = styleElement.getPropertyValue('word-wrap');
 
-  temporaryElement.style.fontFeatureSettings =
-    styleElement.getPropertyValue('font-feature-settings');
-  temporaryElement.style.fontVariantNumeric = styleElement.getPropertyValue('font-variant-numeric');
+  setFontSettings(temporaryElement, styleElement);
 
   temporaryElement.innerHTML = text ?? element.innerHTML;
   return temporaryElement;
@@ -289,26 +293,22 @@ const EllipsisMiddle: React.FC<AsPropsMiddle> = (props) => {
   } = props;
 
   const resizeElement = React.useRef<HTMLDivElement>(null);
-  const [dimension, setDimension] = React.useState<{ fontSize: string; symbolWidth: number }>({
-    fontSize: '14',
-    symbolWidth: 0,
-  });
+  const [symbolWidth, setSymbolWidth] = React.useState(0);
   const blockWidth = useResizeObserver(resizeElement, containerRect).width;
 
   useEnhancedEffect(() => {
     const node = containerRef?.current || resizeElement?.current;
     if (!node) return;
 
+    const styleElement = window.getComputedStyle(node);
     const dateSpan = document.createElement('temporary-block');
-    dateSpan.setAttribute('style', `fontSize: ${dimension.fontSize}px`);
+
+    setFontSettings(dateSpan, styleElement);
     dateSpan.innerHTML = 'a';
     document.body.appendChild(dateSpan);
     const rect = dateSpan.getBoundingClientRect();
 
-    setDimension({
-      fontSize: window.getComputedStyle(node, null).getPropertyValue('font-size'),
-      symbolWidth: rect.width,
-    });
+    setSymbolWidth(rect.width);
     document.body.removeChild(dateSpan);
   }, []);
 
@@ -317,8 +317,8 @@ const EllipsisMiddle: React.FC<AsPropsMiddle> = (props) => {
   const SContainerMiddle = Tooltip;
   const SAdvancedModeContainerMiddle = Tooltip;
   const displayedSymbols = React.useMemo(
-    () => Math.round(blockWidth / dimension.symbolWidth),
-    [blockWidth, dimension.symbolWidth],
+    () => Math.round(blockWidth / symbolWidth),
+    [blockWidth, symbolWidth],
   );
 
   const interaction = text.length > displayedSymbols ? 'hover' : 'none';